Set logging level directly instead of using gflags lib
Circumventing gflags helps since it is not always available or linked correctly. An alternative solution would be to automatically make gflags available through FetchContent, but we would rather keep the number of auto-fetched libraries small to reduce compatibility struggles. Furthermore, our library currently disables the option to install if any dependency is auto-fetched.