Operate directly on node offsets in NodeRef
This is intended to make the NodeRef classes for chunked and linear octrees more similar. It also makes node data accesses faster, by moving the bookkeeping overhead to node child accesses instead. This seems to be reasonable, given that node data accesses tend to be more frequent than child accesses and child accesses also tend to access memory that's further in the cache (so the addition compute latency will likely negligible compared to the cache latency).